requirements.txt
streamlit>=1.36
feedparser>=6.0.11
requests>=2.31
python-dateutil>=2.8.2
